+++ clients/gtest/gesvd_gtest.cpp
@@ -53,31 +53,31 @@ const vector<vector<int>> opt_range = {
     {-1, 0, 0, 2, 2},
     {0, -1, 0, 1, 2},
     {0, 0, -1, 2, 1},
-    {0, -1, 0, 2, 2},
-    {0, 0, -1, 2, 2},
+    {0, 0, 0, 0, 0},
     // normal (valid) samples
     {1, 1, 1, 3, 3},
-    {0, 0, 0, 3, 1},
-    {0, 0, 0, 3, 0},
+    {0, 0, 1, 3, 2},
+    {0, 1, 0, 3, 1},
+    {0, 1, 1, 3, 0},
+    {1, 0, 0, 2, 3},
+    {1, 0, 1, 2, 2},
+    {1, 1, 0, 2, 1},
+    {0, 0, 0, 2, 0},
     {0, 0, 0, 1, 3},
-    {0, 1, 0, 1, 1},
+    {0, 0, 0, 1, 2},
+    {0, 0, 0, 1, 1},
     {0, 0, 0, 1, 0},
     {0, 0, 0, 0, 3},
-    {0, 0, 0, 0, 1},
-    {1, 0, 0, 0, 2},
-    {0, 0, 0, 2, 0},
-    {0, 0, 1, 2, 2}};
+    {0, 0, 0, 0, 2},
+    {0, 0, 0, 0, 1}};
 
 // for daily_lapack tests
-const vector<vector<int>> large_size_range = {{120, 100, 0}, {100, 120, 0}};
+const vector<vector<int>> large_size_range
+    = {{120, 100, 0}, {300, 120, 0}, {300, 120, 1}, {100, 120, 1}, {120, 300, 0}, {120, 300, 1}};
 
 const vector<vector<int>> large_opt_range
     = {{0, 0, 0, 3, 3}, {1, 0, 0, 0, 1}, {0, 1, 0, 1, 0}, {0, 0, 1, 1, 1},
        {0, 0, 0, 3, 0}, {0, 0, 0, 1, 3}, {0, 0, 0, 3, 2}};
-
-// known-bugs tests
-const vector<vector<int>> knownbug_size_range
-    = {{300, 120, 0}, {300, 120, 1}, {120, 300, 0}, {120, 300, 1}};
 
 Arguments gesvd_setup_arguments(gesvd_tuple tup)
 {
@@ -230,8 +230,3 @@ INSTANTIATE_TEST_SUITE_P(daily_lapack,
 // checkin_lapack tests normal execution with small sizes, invalid sizes,
 // quick returns, and corner cases
 INSTANTIATE_TEST_SUITE_P(checkin_lapack, GESVD, Combine(ValuesIn(size_range), ValuesIn(opt_range)));
-
-// known-bugs tests
-INSTANTIATE_TEST_SUITE_P(known_bug_lapack,
-                         GESVD,
-                         Combine(ValuesIn(knownbug_size_range), ValuesIn(large_opt_range)));