+++ CHANGELOG.md
@@ -15,6 +15,16 @@ Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](http
 
 ### Known Issues
 
+### Deprecated
+
+### Removed
+
+### Fixed
+- Thin-SVD failing tests. 
+
+### Known Issues
+
+### Security
 
 ## [rocSOLVER 3.12.0 for ROCm 4.2.0]
 ### Added